+++ src/test/java/Tests/brain/network/CheckPassiveNetworkCategory.java
@@ -1,0 +1,42 @@
+package Tests.brain.network;
+
+import PageObject.BasePage;
+import PageObject.brain.com.ua.BrainHomePage;
+import PageObject.brain.com.ua.BrainNetworkHardwarePage;
+import PageObject.brain.com.ua.BrainOrderPage;
+import PageObject.brain.com.ua.BrainProductPage;
+import Tests.TestInit;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+public class CheckPassiveNetworkCategory extends TestInit {
+
+    @Test
+    public void checkPassiveNetworkCategory() {
+
+        goToSite("https://brain.com.ua/ukr/");
+        BasePage page = new BasePage(driver);
+        page.implicitlyWait();
+
+        BrainHomePage homePage = new BrainHomePage(driver);
+        homePage.getNetworkHardware().click();
+
+        BrainNetworkHardwarePage networkHardwarePage = new BrainNetworkHardwarePage(driver);
+        networkHardwarePage.getPassiveNetwork().click();
+        networkHardwarePage.getSubConnectors().click();
+        homePage.getFirstModel().click();
+
+        BrainProductPage productPage = new BrainProductPage(driver);
+        sleep(5);
+        productPage.getBuyBtn().click();
+        productPage.getCheckOut().click();
+
+        BrainOrderPage orderPage = new BrainOrderPage(driver);
+        orderPage.privat24().click();
+        orderPage.newPost().click();
+        orderPage.recipientData();
+        orderPage.cancelPurchase();
+
+        Assert.assertTrue(orderPage.cartIsEmpty().isDisplayed());
+    }
+}
